+++ .php-cs-fixer.dist.php
@@ -1,8 +1,8 @@
 <?php
 
 $finder = PhpCsFixer\Finder::create()
-    ->in(__DIR__)
-    ->name('MySQLGrid.php');
+    ->in(__DIR__ . '/src')
+    ->name('*.php');
 
 return (new PhpCsFixer\Config())
     ->setRiskyAllowed(false)